+++ apps/docs/app/component/[slug]/layout.client.tsx
@@ -56,16 +56,11 @@ export function ComponentLayoutClient({
     if (pathname.includes('/ux-text')) return 'ux-text'
     if (pathname.includes('/accessibility')) return 'accessibility'
     if (pathname.includes('/code')) return 'code'
+    if (pathname.includes('/faq')) return 'faq'
     return 'overview'
   }
 
   const section = getSection(pathname)
-
-  // const section = pathname.includes('/ux-text')
-  //   ? 'ux-text'
-  //   : pathname.includes('/accessibility')
-  //     ? 'accessibility'
-  //     : 'overview'
 
   if (!isLoaded) return null
 
@@ -163,7 +158,7 @@ export function ComponentLayoutClient({
               overflow="hidden"
               padding="0"
               variant="secondary"
-              border-width="4xs"
+              border-width="5xs"
               onMouseEnter={() => setIsHovered(true)}
               onMouseLeave={() => setIsHovered(false)}
               data-pattern