+++ apps/docs/design/studio/parts/preview/previews/part.preview.shadows.tsx
@@ -1,0 +1,13 @@
+import * as Core from '@sebgroup/green-core/react'
+import { ColorPreviewProps } from '../part.preview.types'
+
+export default function Shadow({ token }: ColorPreviewProps) {
+  return (
+    <Core.GdsCard
+      padding="0"
+      min-height="100%"
+      variant="secondary"
+      box-shadow={token.token}
+    />
+  )
+}